removed outdated tests
